modules GNU e INTEL funcionando

+++ sdumont/2.pre_monan.bash
@@ -15,7 +15,6 @@ export GREEN='\033[1;32m'  # Green
 export RED='\033[1;31m'    # Red
 export NC='\033[0m'        # No Color
 
-
 #----------------------------------
 
 mkdir -p ${DIRMONAN}/logs 
@@ -24,41 +23,36 @@ mkdir -p ${DIRMONAN}/tar
   RES=40962
   RES=1024002
 
-function firstPart(){ # data part 
+function dataPhase(){ # data part 
   echo ----------------    em function firstPart
   echo -e  "${GREEN}==>${NC} Copying and decompressing testcase data... \n"
   echo -e  "${GREEN}==>${NC}  and submit init_atmospere ... \n"
   # Temporariamente, enquanto desenv:----------------------------------------------v
-  #wget ${FTPADD}/${DIRDADOS}/MONAN_testcase_v1.0.tgz 
+  wget ${FTPADD}/${DIRDADOS}/MONAN_testcase_v1.0.tgz 
   #CR: TODO: verificar se o wget baixou corretamente o dado antes de destargear:
-  #tar -xzf ./MONAN_testcase_v1.0.tgz -C ${DIRroot}
-  comando="time tar -xzf /tmp/${DIRDADOS}/MONAN_testcase_v1.0.tgz -C ${DIRroot}"
-  #echo $comando; eval $comando
-  comando="time tar -xzf /tmp/${DIRDADOS}/MONAN_data_v1.0.tgz     -C ${DIRMONAN}"
-  #echo $comando; eval $comando
   comando="cp -f  ../../../rpsouto/sequana/projetos/monan/ungrib/ungrib.exe MONAN/exec/ungrib.exe"
-  #echo $comando; eval $comando
-  # cp  ../../../rpsouto/sequana/projetos/monan/ungrib/ungrib.exe MONAN/exec/ungrib_SD.exe
-  #if [ ! -s ${DIRDADOS}/MONAN_testcase_v1.0.tgz ] 
-  #then
-  #   echo "dado nao existe no /tmp/${DIRDADOS}/MONAN_testcase_v1.0.tgz"
-  #   exit
-  #fi
-  #tar -xzf ${DIRDADOS}/MONAN_testcase_v1.0.tgz -C ${DIRroot}
+  echo $comando; eval $comando
+  cp  ../../../rpsouto/sequana/projetos/monan/ungrib/ungrib.exe MONAN/exec/ungrib_SD.exe
+  if [ ! -s MONAN_testcase_v1.0.tgz ] 
+  then
+     echo "dado nao existe no MONAN_testcase_v1.0.tgz"
+     exit
+  fi
+  tar -xzf MONAN_testcase_v1.0.tgz -C ${DIRroot}
   # Temporariamente, enquanto desenv:----------------------------------------------^
 
   echo -e  "${GREEN}==>${NC} Copyings scripts from MONAN_ori to MONAN testcase script folders... \n"
-  cp -rf ${DIRMONAN_ORI}/testcase/scripts/* ${DIRMONAN}/testcase/scripts/
+  comando="cp -rf ${DIRMONAN_ORI}/testcase/scripts/* ${DIRMONAN}/testcase/scripts/"
+  echo "em $(basename $0), $comando"; eval $comando
 
   echo -e  "${GREEN}==>${NC} Copying and decompressing all data for preprocessing... \n"
   echo -e  "${GREEN}==>${NC} It may take several minutes...\n"
   #CR: TODO: inserir opcao "timestamping" no wget:  baixa o arq somente sei verificar que o arq ja existe no dir local. (testar)
 
   # Temporariamente, enquanto desenv:----------------------------------------------v
-  #wget ${FTPADD}/${DIRDADOS}/MONAN_data_v1.0.tgz
+  wget ${FTPADD}/${DIRDADOS}/MONAN_data_v1.0.tgz
   #CR: TODO: verificar se o wget baixou corretamente o dado antes de destargear:
   #CR: TODO: incluir o dir MONAN dentro do tar MONAN_data_v1.0.tgz para fim de padronizacao.
-  #tar -xzf ${DIRDADOS}/MONAN_data_v1.0.tgz -C ${DIRMONAN}
   if [ ! -s ${DIRDADOS_LOCAL}/MONAN_data_v1.0.tgz ] 
   then
      echo "dado nao existe no ${DIRDADOS_LOCAL}/MONAN_data_v1.0.tgz"
@@ -83,50 +77,95 @@ function firstPart(){ # data part
      done
      date
   fi
-} #  function firstPart
+} #  function dataPhase
 
-function firstPartB (){
+function 40962dataA(){
+
+  comando="cp /scratch/cenapadrjsd/eduardo.garcia2/monan/$RES/x1.$RES.grid.nc                     ${DIRroot}/MONAN/testcase/data/meshes/"
+  echo $comando; eval $comando
+
+  comando="cp /scratch/cenapadrjsd/eduardo.garcia2/monan/$RES/x1.$RES.graph.info.part.$numNucleos ${DIRroot}/MONAN/testcase/namelist/"
+  echo $comando; eval $comando
+
+} # function 40962dataA(){
+
+function 40962dataB(){
+
+  comando="cp /scratch/cenapadrjsd/eduardo.garcia2/monan/$RES/static/namelist.init_atmosphere  ${DIRroot}/MONAN/testcase/runs/ERA5/static/; "
+  #echo $comando; eval $comando
+
+  comando="cp /scratch/cenapadrjsd/eduardo.garcia2/monan/$RES/static/streams.init_atmosphere   ${DIRroot}/MONAN/testcase/runs/ERA5/static/; "
+  #echo $comando; eval $comando
+
+  comando="cp /scratch/cenapadrjsd/eduardo.garcia2/monan/$RES/namelist.init_atmosphere  ${DIRroot}/MONAN/testcase/runs/ERA5/2021010100/; "
+  echo $comando; eval $comando
+
+  comando="cp /scratch/cenapadrjsd/eduardo.garcia2/monan/$RES/streams.init_atmosphere   ${DIRroot}/MONAN/testcase/runs/ERA5/2021010100/; "
+  echo $comando; eval $comando
+
+  comando="cp /scratch/cenapadrjsd/eduardo.garcia2/monan/$RES/namelist.atmosphere       ${DIRroot}/MONAN/testcase/runs/ERA5/2021010100/; "
+  echo $comando; eval $comando
+
+  comando="cp /scratch/cenapadrjsd/eduardo.garcia2/monan/$RES/streams.atmosphere        ${DIRroot}/MONAN/testcase/runs/ERA5/2021010100/; "
+  echo $comando; eval $comando
+
+} # function 40962dataB(){
+
+function static (){
 
   # Temporariamente, enquanto desenv:----------------------------------------------^
   echo -e  "${GREEN}==>${NC} Creating make_static.sh for submiting static init_atmosphere...\n"
-  echo -e  "${GREEN}==>${NC}    and data: stream.init_atmosphere namelist.init_atmosphere \n"
+  echo -e  "${GREEN}==>${NC}    and data: streams.init_atmosphere namelist.init_atmosphere \n"
   cd ${DIRMONAN}/testcase/scripts
+  echo PWD="$(pwd)"
   comando="${DIRMONAN}/testcase/scripts/static.sh ERA5 $RES"
+  echo  $comando ; eval $comando
+  comando="grep \"x1\" /scratch/cenapadrjsd/eduardo.garcia2/MONAN-scripts/sdumont/MONAN/testcase/runs/ERA5/static/*.init_atmosphere*"
   echo $comando ; eval $comando
 
+  #read -p "esperando um ok!"
   echo -e  "${GREEN}==>${NC} Submiting sbatch make_static.sh...\n"
   cd ${DIRMONAN}/testcase/runs/ERA5/static
-  #grep 40 make_static.sh
-  grep "x1.*" /scratch/cenapadrjsd/eduardo.garcia2/MONAN-scripts/sdumont/MONAN/testcase/runs/ERA5/static/*.init_atmosphere
+  pwd
   date
-  comando="sbatch --wait                                make_static.sh"
-  echo $comando ;# read -p "arguardando um ok!"
-  eval $comando
+  comando="sbatch --wait make_static.sh"
+  echo $comando; eval $comando
   date
 
+#  return
   if [ ! -e x1.$RES.static.nc ]; then
     echo -e  "\n${RED}==>${NC} ***** ATTENTION *****\n"	
     echo -e  "${RED}==>${NC} Static phase fails ! Check logs at ${DIRMONAN}/testcase/runs/ERA5/static/logs . Exiting script. \n"
     exit 1
+  else
+    echo -e  "${GREEN}==>${NC} Static phase finished ! Check logs at ${DIRMONAN}/testcase/runs/ERA5/static/logs.\n"
   fi
 
-} #  end function firstPartB (){
+} #  end function static (){
 
 function criarDataAndSLURMScripts (){
   echo -e  "${GREEN}==>${NC} Creating submition script to ungrib,          degrib_exe.sh \n"
   echo -e  "${GREEN}==>${NC} Creating submition script to init_atmosphere, Init_atmos.sh \n"
-  echo -e  "${GREEN}==>${NC}    and data : stream.init_atmosphere namelist.init_atmosphere \n"
+  echo -e  "${GREEN}==>${NC}    and data : streams.init_atmosphere namelist.init_atmosphere \n"
   echo -e  "${GREEN}==>${NC} Creating submition script to atmosphere,      monan_exe.sh  \n"
-  echo -e  "${GREEN}==>${NC}    and data : stream.atmosphere namelist.atmosphere \n"
+  echo -e  "${GREEN}==>${NC}    and data : streams.atmosphere namelist.atmosphere \n"
   cp -rf ${DIRMONAN_ORI}/testcase/scripts/* ${DIRMONAN}/testcase/scripts/
   cd ${DIRMONAN}/testcase/scripts
+
+  if [ $RES -eq "40962" ] ; then
+     sed -i '/RES=/a RES=40962' ${DIRMONAN}/testcase/scripts/run_monan.bash;
+  fi
+
   comando="${DIRMONAN}/testcase/scripts/run_monan.bash ERA5 2021010100"
-  echo $comando; eval $comando
+  echo "em $(basename $0), $comando"; eval $comando
+
+  if [ $RES -eq "40962" ] ; then 40962dataB; fi
+  echo "fim de criarDataAndSLURMScripts"
+
 }
 
-function secondPart(){ # make_degrib
-  echo ----------------    em function secondPart
-  criarDataAndSLURMScripts
+function degrib(){ # make_degrib
+  echo ----------------    em function degrib
   HOME=$SCRATCH
 
   echo ----------------    submiting script : degrib_exe.sh  
@@ -138,10 +177,12 @@ function secondPart(){ # make_degrib
   cd ${DIRMONAN}/testcase/runs/ERA5/2021010100/wpsprd/
 
   echo -e  "${GREEN}==>${NC} Submiting degrib_exe.sh...\n"
+  pwd
   date
   comando="sbatch --wait degrib_exe.sh"
-  echo $comando ; eval $comando
+  echo $comando ;  eval $comando
   date
+#  return
 
   files_ungrib=("LSM:1979-01-01_00" "GEO:1979-01-01_00" "FILE:2021-01-01_00" "FILE2:2021-01-01_00" "FILE3:2021-01-01_00")
   for file in "${files_ungrib[@]}"; do
@@ -152,46 +193,44 @@ function secondPart(){ # make_degrib
       exit  1
     fi
   done
-} # function secondPart(){
+  echo -e  "${GREEN}==>${NC} Degrib finished ! Check logs at ${DIRMONAN}/testcase/runs/ERA5/2021010100/wpsprd/. \n"
+} # function degrib(){
 
-function thirdPart(){ # make_initatmos
-  echo ----------------    em function thirdPart
+function initAtmos(){ # make_init_atmos
+  echo ----------------    em function initAtmos
   echo -e  "${GREEN}==>${NC} Submiting script InitAtmos_exe.sh...\n"
   cd ${DIRMONAN}/testcase/runs/ERA5/2021010100
+  pwd
   date
-  comando="sbatch --wait -p sequana_cpu_dev -t 00:20:00 InitAtmos_exe.sh"
-  comando="sbatch --wait                                InitAtmos_exe.sh"
-  echo $comando; eval $comando
+  comando="sbatch --wait InitAtmos_exe.sh"
+  echo $comando;  eval $comando
   date
+#  return
   if [ ! -e x1.$RES.init.nc ]; then
     echo -e  "\n${RED}==>${NC} ***** ATTENTION *****\n"	
     echo -e  "${RED}==>${NC} Init Atmosphere phase fails ! Check logs at ${DIRMONAN}/testcase/runs/ERA5/2021010100/logs . Exiting script.\n"
     exit -1
+  else
+    echo -e  "${GREEN}==>${NC} Init Atmosphere phase finished ! Check logs at ${DIRMONAN}/testcase/runs/ERA5/2021010100/logs.\n"
   fi
   echo -e  "${GREEN}==>${NC} Script ${0} completed. \n"
-} # function thirdPart
+} # function initAtmos
 
-function 40962data(){
-
-RESL=40962
-cp /scratch/cenapadrjsd/eduardo.garcia2/monan/$RESL/namelist.init_atmosphere  /scratch/cenapadrjsd/eduardo.garcia2/MONAN-scripts/sdumont//MONAN/testcase/runs/ERA5/2021010100/; 
-cp /scratch/cenapadrjsd/eduardo.garcia2/monan/$RESL/streams.init_atmosphere   /scratch/cenapadrjsd/eduardo.garcia2/MONAN-scripts/sdumont//MONAN/testcase/runs/ERA5/2021010100/; 
-cp /scratch/cenapadrjsd/eduardo.garcia2/monan/$RESL/namelist.atmosphere  /scratch/cenapadrjsd/eduardo.garcia2/MONAN-scripts/sdumont//MONAN/testcase/runs/ERA5/2021010100/; 
-cp /scratch/cenapadrjsd/eduardo.garcia2/monan/$RESL/streams.atmosphere   /scratch/cenapadrjsd/eduardo.garcia2/MONAN-scripts/sdumont//MONAN/testcase/runs/ERA5/2021010100/; 
-
-}
-
+#shift
 source ./load_monan_app_modules.sh $COMPILER
 
-firstPart  # testcase data phase, originals scripts  + ungrib
+#dataPhase  # testcase data phase, originals scripts  + ungrib
 
-firstPartB # submit make_static.sh -> x1.$RES.static.nc
+if [ $RES -eq "40962" ] ; then 40962dataA; fi
 
-secondPart #  submit degrid_exe.sh  -> directory wpsprd
-#40962data
-grep "x1.*" /scratch/cenapadrjsd/eduardo.garcia2/MONAN-scripts/sdumont/MONAN/testcase/runs/ERA5/2021010100/*.init_atmosphere
+static #  static.sh e submit make_static.sh -> x1.$RES.static.nc
 
-#read -p "waiting for a command"
-thirdPart  # submit Init_atmos_exe.sh -> x1.$RES.init.nc
-pwd
-grep "x1.*" /scratch/cenapadrjsd/eduardo.garcia2/MONAN-scripts/sdumont/MONAN/testcase/runs/ERA5/2021010100/*.atmosphere
+criarDataAndSLURMScripts  # run_monan.bash;
+comando="grep \"x1\" /scratch/cenapadrjsd/eduardo.garcia2/MONAN-scripts/sdumont/MONAN/testcase/runs/ERA5/2021010100/*.init_atmosphere"
+echo $comando ; eval $comando
+comando="grep \"x1\" /scratch/cenapadrjsd/eduardo.garcia2/MONAN-scripts/sdumont/MONAN/testcase/runs/ERA5/2021010100/*.atmosphere"
+echo $comando ; eval $comando
+
+degrib # submit degrid_exe.sh  -> directory wpsprd
+
+initAtmos  # submit Init_atmos_exe.sh -> x1.$RES.init.nc
